Add SEO metadata for key public pages
