+++ vitest.config.ts
@@ -1,10 +1,11 @@
-import { defineConfig } from "vitest/config";
+import { configDefaults, defineConfig } from "vitest/config";
 import path from "node:path";
 
 export default defineConfig({
   test: {
     environment: "node",
     globals: true,
+    exclude: [...configDefaults.exclude, "realtime-server/**"],
   },
   resolve: {
     alias: {
